Web Summarizer Agent
Web Summarizer Agent
The readme is AI generated.
README instructions from other branch are relevant for packaging this as docker image, and pushing to AWS lambda/testing locally with the provided test UI.

This implementation is NOT yet ready, but it works on Lambda.
It's not ready because it needs one small tweak, to change the Embedding provider. The embedding I set up in this code works, but is completely impractical in a realistic scenario, because it makes the docker image over 10GB, due to pytorch (huggingface transformer dependency)

So, it only works locally, until the embedding provider is changed from huggingface, and the dependencies are updated to remove huggingface/pytorch and replace it(anything is better than locally hosting it in the docker container)